requirementst.txt
numpy
pandas
scikit-learn
mlflow
dvc
dvc_s3
boto3
pyyaml
